+++ src/classes/providers/codelens/jscodelens/JSCodeLensProvider.ts
@@ -6,28 +6,24 @@ import { OverridenMethodCodeLensGenerator } from "./strategies/OverridenMethodCo
 import { ReferenceCodeLensGenerator } from "./strategies/ReferenceCodeLensGenerator";
 
 export class JSCodeLensProvider {
-	static getCodeLenses(document: vscode.TextDocument): Promise<vscode.CodeLens[]> {
-		return new Promise(resolve => {
-			// setTimeout(() => {
-			UIClassFactory.setNewContentForClassUsingDocument(document);
-			let codeLenses: vscode.CodeLens[] = [];
+	static async getCodeLenses(document: vscode.TextDocument) {
+		// setTimeout(() => {
+		UIClassFactory.setNewContentForClassUsingDocument(document);
+		const codeLenses: vscode.CodeLens[] = [];
 
-			const aStrategies = [
-				InternalizationTextCodeLenseGenerator,
-				OverridenMethodCodeLensGenerator,
-				EventHandlerCodeLensGenerator,
-				ReferenceCodeLensGenerator
-			];
-			// console.time("Code lens");
-			aStrategies.forEach(Strategy => {
-				const strategy = new Strategy();
-				codeLenses = strategy.getCodeLenses(document).concat(codeLenses);
-			});
-			// console.timeEnd("Code lens");
-
-			resolve(codeLenses);
-			// copy(JSON.stringify(codeLenses.map(codeLens => codeLens.command.title)))
-			// }, 0);
+		const aStrategies = [
+			InternalizationTextCodeLenseGenerator,
+			OverridenMethodCodeLensGenerator,
+			EventHandlerCodeLensGenerator,
+			ReferenceCodeLensGenerator
+		];
+		// console.time("Code lens");
+		aStrategies.forEach(Strategy => {
+			const strategy = new Strategy();
+			codeLenses.push(...strategy.getCodeLenses(document));
 		});
+		// console.timeEnd("Code lens");
+
+		return codeLenses;
 	}
 }