+++ src/classes/providers/diagnostics/js/jslinter/parts/WrongFieldMethodLinter.ts
@@ -11,72 +11,40 @@ import { RangeAdapter } from "../../../../../adapters/vscode/RangeAdapter";
 export class WrongFieldMethodLinter extends Linter {
 	protected className = "WrongFieldMethodLinter";
 	public static timePerChar = 0;
-	private _timeSpent = 0;
-	async _getErrors(document: vscode.TextDocument) {
+	_getErrors(document: vscode.TextDocument): IError[] {
 		let errors: IError[] = [];
 
 		if (vscode.workspace.getConfiguration("ui5.plugin").get("useWrongFieldMethodLinter")) {
 			// console.time("WrongFieldMethodLinter");
-			errors = await this._getLintingErrors(document);
+			const start = new Date().getTime();
+			errors = this._getLintingErrors(document);
+			const end = new Date().getTime();
+			WrongFieldMethodLinter.timePerChar = (end - start) / document.getText().length;
 			// console.timeEnd("WrongFieldMethodLinter");
-			WrongFieldMethodLinter.timePerChar = this._timeSpent / document.getText().length;
 		}
 
 		return errors;
 	}
-
-	private async _getLintingErrors(document: vscode.TextDocument) {
+	private _getLintingErrors(document: vscode.TextDocument): IError[] {
 		let errors: IError[] = [];
-		const approximateTime = WrongFieldMethodLinter.timePerChar * document.getText().length;
-		let partQuantity = 1;
-		const pauseBetweenParts = 100;
-		if (approximateTime > pauseBetweenParts) {
-			partQuantity = Math.ceil(approximateTime / pauseBetweenParts);
-		}
+
 		const currentClassName = FileReader.getClassNameFromPath(document.fileName);
 		if (currentClassName) {
 			const UIClass = <CustomUIClass>UIClassFactory.getUIClass(currentClassName);
 			const acornMethods = UIClass.acornMethodsAndFields.filter(fieldOrMethod => fieldOrMethod.value.type === "FunctionExpression").map((node: any) => node.value.body);
 
-			if (partQuantity > acornMethods.length) {
-				partQuantity = acornMethods.length;
-			}
-			let parts = [];
-			const methodContainerLength = Math.ceil(acornMethods.length / partQuantity);
-
-			for (let index = 0; index < partQuantity; index++) {
-				parts.push(acornMethods.slice(index * methodContainerLength, index * methodContainerLength + methodContainerLength));
-			}
-			parts = parts.filter(part => part.length > 0);
-
-			for (const part of parts) {
-				errors = await this._getLintingErrorsForMethods(part, UIClass, document, 0);
-			}
+			acornMethods.forEach((method: any) => {
+				if (method.body) {
+					method.body.forEach((node: any) => {
+						const validationErrors = this._getErrorsForExpression(node, UIClass, document);
+						errors = errors.concat(validationErrors);
+					});
+				}
+			});
 
 		}
 
 		return errors;
-	}
-
-	private _getLintingErrorsForMethods(acornMethods: any[], UIClass: CustomUIClass, document: vscode.TextDocument, timeout: number): Promise<IError[]> {
-		return new Promise((resolve) => {
-			setTimeout(() => {
-				const start = new Date().getTime();
-				let errors: IError[] = [];
-				acornMethods.forEach((acornMethod: any) => {
-					if (acornMethod.body) {
-						acornMethod.body.forEach((node: any) => {
-							const validationErrors = this._getErrorsForExpression(node, UIClass, document);
-							errors = errors.concat(validationErrors);
-						});
-					}
-				});
-
-				const end = new Date().getTime();
-				this._timeSpent += end - start;
-				resolve(errors);
-			}, timeout);
-		});
 	}
 
 	private _getErrorsForExpression(node: any, UIClass: CustomUIClass, document: vscode.TextDocument, errors: IError[] = [], droppedNodes: any[] = [], errorNodes: any[] = []) {
@@ -152,8 +120,7 @@ export class WrongFieldMethodLinter extends Linter {
 							} else {
 								const member = singleFieldsAndMethods.fields.find(field => field.name === nextNodeName) || singleFieldsAndMethods.methods.find(method => method.name === nextNodeName);
 								const isIgnored = !!(<UI5Ignoreable>member)?.ui5ignored;
-								const ignoreAccessLevelModifiers = vscode.workspace.getConfiguration("ui5.plugin").get("ignoreAccessLevelModifiers");
-								if (!ignoreAccessLevelModifiers && !isIgnored) {
+								if (!isIgnored) {
 									let sErrorMessage = "";
 									if (member?.visibility === "protected") {
 										const currentDocumentClassName = FileReader.getClassNameFromPath(document.fileName);