+++ src/test/project/src/util/interfacetest/interfaces/Interface2.js
@@ -1,0 +1,11 @@
+sap.ui.define([
+], function(
+) {
+	"use strict";
+
+	return {
+		interface2Method1() {},
+		interface2Method2() {},
+		interface2Method3() {}
+	};
+});